refactor(synchronizer): cleanup code
