+++ packages/synchronizer/src/createDefaultMonokleAuthenticator.ts
