+++ packages/synchronizer/src/index.ts
@@ -1,7 +1,12 @@
 export * from './handlers/apiHandler';
 export * from './handlers/deviceFlowHandler';
+export * from './handlers/gitHandler';
 export * from './handlers/storageHandler';
+export * from './handlers/storageHandlerAuth';
+export * from './handlers/storageHandlerPolicy';
 
 export * from './utils/authenticator';
+export * from './utils/synchronizer';
 
-export * from './createDefaultAuthenticator';
+export * from './createDefaultMonokleAuthenticator';
+export * from './createDefaultMonokleSynchronizer';
